added ItemsUtil to handle articleList operation.
